+++ trunk/melting-java/src/melting/approximativeMethods/WetmurDNARNA91.java
@@ -13,12 +13,12 @@ public class WetmurDNARNA91 extends Wetmur91 {
 	private static String temperatureEquation = "Tm = 67 + 16.6 * log10(Na / (1.0 + 0.7 * Na)) + 0.8 * percentGC - 500 / duplexLength - percentMismatching";
 	
 	public ThermoResult CalculateThermodynamics() {
-		int percentGC = this.environment.getSequences().calculatePercentGC();
-		int percentMismatching = this.environment.getSequences().getPercentMismatching();
+		double percentGC = this.environment.getSequences().calculatePercentGC();
+		double percentMismatching = this.environment.getSequences().getPercentMismatching();
 		int duplexLength = this.environment.getSequences().getDuplexLength();
 		double Tm = super.CalculateThermodynamics().getTm();
 		
-		Tm = 67 + 16.6 * Math.log10(this.environment.getNa() / (1.0 + 0.7 *this.environment.getNa())) + 0.8 * percentGC - 500 / duplexLength - percentMismatching;
+		Tm = 67.0 + 16.6 * Math.log10(this.environment.getNa() / (1.0 + 0.7 *this.environment.getNa())) + 0.8 * percentGC - 500.0 / duplexLength - percentMismatching;
 
 		this.environment.setResult(Tm);
